requirements.txt
docplex
emllib
matplotlib
numpy
pandas
pydoe
scikit-learn
seaborn
tensorflow